+++ snooplyze/notifier/telegram_notifier.py
@@ -1,5 +1,5 @@
 import requests
-from notifier import Notifier
+from snooplyze.notifier import Notifier
 
 
 class TelegramNotifier(Notifier):