+++ snooplyze/persistence/duckdb_persistence_engine.py
@@ -1,7 +1,7 @@
 import duckdb
 
 from .base_persistence_engine import PersistenceEngine
-from page import PageContent
+from snooplyze.page import PageContent